requirements/base.txt
# Bleeding edge Django... really?
django>=1.10.5,<1.11

# Configuration
django-environ==0.4.1



# Forms
django-braces==1.11.0
django-crispy-forms==1.6.1

# Models
django-model-utils==2.6.1

# Images
Pillow>=4.0.0

# For user registration, either via email or social
# Well-built with regular release cycles!
django-allauth>=0.30.0

# On Windows, you must download/install psycopg2 manually
# from http://www.lfd.uci.edu/~gohlke/pythonlibs/#psycopg


# Unicode slugification
awesome-slugify==1.6.5

# Time zones support
pytz==2016.10

# Redis support
django-redis==4.7.0
redis>=2.10.5


#celery==3.1.24



rcssmin==1.0.6
django-compressor==2.1.1


# Your custom requirements go here
psycopg2==2.6.2
django-imagekit==4.0
django-bootstrap3>=8.2.2
